chatbot_service/requirements.txt
# Web Framework
fastapi==0.103.2
uvicorn==0.23.2

# Database
pymysql==1.1.0

# Cache
redis==5.0.1

# AI & Machine Learning
transformers==4.35.2
torch==2.1.0
sentencepiece
protobuf
numpy==1.24.3

# LangChain
langchain>=0.1.0
langchain-community>=0.0.13
langchain-pinecone>=0.0.1
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0

# Vector Database
pinecone-client==3.2.2
pinecone-text>=0.5.0

# PDF Processing
pypdf==4.2.0

# Embeddings
sentence-transformers==2.7.0

# Google AI
google-generativeai

# Utilities
python-dotenv==1.0.0
requests==2.31.0
